chore(execution): separate tests and test utils to files
