+++ crates/papyrus_gateway/src/v0_4_0/mod.rs
@@ -3,6 +3,8 @@ pub mod block;
 pub mod broadcasted_transaction;
 pub mod deprecated_contract_class;
 pub mod error;
+#[cfg(test)]
+mod execution_test;
 pub mod state;
 pub mod transaction;
 pub mod write_api_error;
